first 50 fibonacci numbers corrected

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,8 +6,8 @@
  */
 int main(void)
 {
-	int f1 = 1;
-	int f2 = 2;
+	long int f1 = 1;
+	long int f2 = 2;
 	long int F;
 	int i;
